file/directory rename to be in line with project name change. Parsing code updated to no longer use manually copied categorical maps in /uns, instead readinf directly from /obs and /var.
