+++ src/main/kotlin/graphics/scenery/corvo/Util.kt
@@ -1,0 +1,5 @@
+package graphics.scenery.corvo
+
+typealias CellIndex = Int
+typealias GeneIndex = Int
+typealias Type = String
