--- src/main/kotlin/graphics/scenery/xtradimensionvr/Util.kt
@@ -1,5 +1,0 @@
-package graphics.scenery.xtradimensionvr
-
-typealias CellIndex = Int
-typealias GeneIndex = Int
-typealias Type = String
